# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/admin/notifications": {
      "get": {
        "operationId": "Notifications_list",
        "summary": "",
        "description": "Outbox rows, newest first.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "SENT",
                "FAILED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "e.g. order_confirmed, event_changed (docs/notifications.md)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/notifications/summary": {
      "get": {
        "operationId": "Notifications_summary",
        "summary": "",
        "description": "Counts by status, for a quick health check.\n\n**Roles:** ADMIN",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/notifications/{id}/retry": {
      "post": {
        "operationId": "Notifications_retry",
        "summary": "",
        "description": "Queue a FAILED or CANCELLED message again (e.g. after fixing the mail settings).\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/notifications/run": {
      "post": {
        "operationId": "Notifications_run",
        "summary": "",
        "description": "Send everything that's due now instead of waiting for the next poll.\n\n**Roles:** ADMIN",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/notifications/scan-reminders": {
      "post": {
        "operationId": "Notifications_scanReminders",
        "summary": "",
        "description": "Queue due event reminders now instead of waiting for the next scan.\n\n**Roles:** ADMIN",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/organizers": {
      "get": {
        "operationId": "OrganizersAdmin_list",
        "summary": "",
        "description": "Organizers with their approval, trust level and effective permissions.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "verificationStatus",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrganizerVerificationStatus"
            }
          },
          {
            "name": "trustLevel",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrganizerTrustLevel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Organizers Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/organizers/{id}": {
      "get": {
        "operationId": "OrganizersAdmin_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Organizers Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      },
      "patch": {
        "operationId": "OrganizersAdmin_update",
        "summary": "",
        "description": "Approve/suspend, set the trust level, override single permissions or limits.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizerTrustDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Organizers Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/events/review": {
      "get": {
        "operationId": "OrganizersAdmin_reviewQueue",
        "summary": "",
        "description": "Events waiting for review, oldest first.\n\n**Roles:** ADMIN",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Organizers Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/events/{id}/approve": {
      "post": {
        "operationId": "OrganizersAdmin_approve",
        "summary": "",
        "description": "Approve: the event goes on sale and the organizer is emailed.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Organizers Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/events/{id}/reject": {
      "post": {
        "operationId": "OrganizersAdmin_reject",
        "summary": "",
        "description": "Send back to draft with a note telling the organizer what to change.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Organizers Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Event Ticketing Platform API",
    "description": "REST API shared by the web app and the organizer/staff mobile apps. Amounts are integer minor units (butut, GMD). Sign in with POST /api/v1/auth/login, then send the access token as `Authorization: Bearer <token>`. See docs/api.md.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OrganizerVerificationStatus": {
        "type": "string",
        "description": "APPROVED lets them publish; SUSPENDED stops all their ticket sales at once.",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "SUSPENDED"
        ]
      },
      "OrganizerTrustLevel": {
        "type": "string",
        "enum": [
          "NEW",
          "TRUSTED"
        ]
      },
      "UpdateOrganizerTrustDto": {
        "type": "object",
        "properties": {
          "verificationStatus": {
            "$ref": "#/components/schemas/OrganizerVerificationStatus"
          },
          "trustLevel": {
            "$ref": "#/components/schemas/OrganizerTrustLevel"
          },
          "requireEventReview": {
            "type": "boolean",
            "nullable": true,
            "description": "Override; null = level default"
          },
          "canConfirmBankTransfers": {
            "type": "boolean",
            "nullable": true,
            "description": "Override; null = level default"
          },
          "canHandleCancellationRefunds": {
            "type": "boolean",
            "nullable": true,
            "description": "Override; null = level default"
          },
          "customLimits": {
            "type": "boolean",
            "description": "true: use maxTicketsPerEvent / maxTicketPrice below instead of the level's limits"
          },
          "maxTicketsPerEvent": {
            "type": "number",
            "nullable": true,
            "description": "With customLimits; null = no limit",
            "minimum": 1,
            "maximum": 1000000
          },
          "maxTicketPrice": {
            "type": "number",
            "nullable": true,
            "description": "With customLimits, minor units; null = no limit",
            "minimum": 0,
            "maximum": 1000000000
          },
          "verifiedBadge": {
            "type": "boolean",
            "description": "Verified badge (blue tick) shown next to their name: for official organizers whose identity the platform has confirmed. Needs an approved account."
          },
          "payoutAdvancePercent": {
            "type": "number",
            "description": "Share (0-100) of an upcoming event's earnings they may be paid before it happens. 0 = only after the event (docs/payouts.md).",
            "minimum": 0,
            "maximum": 100
          },
          "note": {
            "type": "string",
            "description": "Private admin note (why)",
            "maxLength": 1000
          }
        }
      },
      "ReviewDecisionDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "Required when sending back: what the organizer should change",
            "maxLength": 1000
          }
        }
      },
...
    }
  }
}